refactor: share code to remove hyperHTML comments
